+++ src/mo_interpreter/interpret.ml
@@ -1002,7 +1002,11 @@ and interpret_block env decs ro (k : V.value V.cont) =
 and declare_dec dec : val_env =
   match dec.it with
   | ExpD _
-  | TypD _ -> V.Env.empty
+  | TypD _
+  | MixinD (_) -> V.Env.empty
+  | IncludeD _ ->
+     (* TODO support mixins in the interpreter *)
+    assert false
   | LetD (pat, _, _) -> declare_pat pat
   | VarD (id, _) -> declare_id id
   | ClassD (_eo, _, _, id, _, _, _, _, _) -> declare_id {id with note = ()}
@@ -1035,6 +1039,15 @@ and interpret_dec env dec (k : V.value V.cont) =
     )
   | TypD _ ->
     k V.unit
+  | MixinD _ -> k V.unit
+  | IncludeD (_, _arg, _note) ->
+     (* TODO
+        - evaluate arg and bind it against note.pat
+        - define note.imports from mixin as local lets
+        - declare/eval note.decs
+        - Do we need to recreate the renaming pass?
+      *)
+     trap dec.at "Mixins are not yet supported in the interpreter"
   | ClassD (_eo, shared_pat, obj_sort, id, _typbinds, pat, _typ_opt, id', dec_fields) ->
     (* NB: we ignore the migration expression _eo *)
     let f = interpret_func env id.it shared_pat pat (fun env' k' ->